Added new maze museum background, night guard sprite, and updated start screen

+++ GEM_Game/inventory.cpp
@@ -1,4 +1,4 @@
-#include "Inventory.h"
+#include "inventory.h"
 #include <iostream>
 using namespace std;
 
@@ -7,7 +7,7 @@ void Inventory::addItem(Item* item) {
 }
 
 void Inventory::showItems() {
-    for (int i = 0; i < items.size(); i++) {
+    for (size_t i = 0; i < items.size(); i++) {
         cout << items[i]->getName() << endl;
     }
 }
@@ -15,7 +15,7 @@ void Inventory::showItems() {
 int Inventory::countItem(string name) {
     int count = 0;
 
-    for (int i = 0; i < items.size(); i++) {
+    for (size_t i = 0; i < items.size(); i++) {
         if (items[i]->getName() == name) {
             count++;
         }
